refactor: split audit.go into core + control + route (#540)
audit.go was 954 lines — over the project's ~300-line guideline.
Split per-event/per-category control methods and per-output route
validation/management into two new sibling files. Pure refactor;
no behaviour change, no public API change.

Kept in audit.go (801 lines):
- Pools, constants
- Auditor struct
- Construction (New + helpers)
- Logger API
- Core event pipeline (AuditEvent, AuditEventContext, validate,
  enqueue, deliverSync, dropOnBufferFull, recordCtxCancelDrop)
- Shutdown (Close, closeOutputs, waitForDrain)
- Handle / MustHandle
- Field pool helpers (returnFieldsToPool, copyFieldsWithDefaults)

Moved to control.go (96 lines):
- Auditor.EnableCategory, DisableCategory, EnableEvent, DisableEvent
- Inline rationale for the fmt.Errorf allocation trade-off (#509,
  C-23) — control-plane allocation is acceptable, hot path has
  separate optimisation.

Moved to route.go (102 lines):
- Auditor.validateOutputRoutes, validateExcludeLabels
- Auditor.SetOutputRoute, ClearOutputRoute, OutputRoute

audit.go remains over the 300-line target; further splitting risks
separating tightly-coupled hot-path code (core pipeline, drain,
validate, enqueue). Documented as accepted exception.

Verification:
- go build ./... → clean
- go test -race -count=1 . → green (180s)

Refs #540.
