+++ audit.go
@@ -695,163 +695,10 @@ func (a *Auditor) waitForDrain() {
 	}
 }
 
-// validateOutputRoutes checks all per-output event routes and
-// sensitivity exclusion labels against the taxonomy.
-func (a *Auditor) validateOutputRoutes() error {
-	for _, oe := range a.entries {
-		route := oe.route.Load()
-		if route != nil {
-			if err := ValidateEventRoute(route, a.taxonomy); err != nil {
-				return fmt.Errorf("audit: output %q: %w", oe.output.Name(), err)
-			}
-		}
-		if err := a.validateExcludeLabels(oe); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// validateExcludeLabels checks that all exclude_labels on an output
-// reference labels defined in the taxonomy's sensitivity config.
-func (a *Auditor) validateExcludeLabels(oe *outputEntry) error {
-	if len(oe.excludedLabels) == 0 {
-		return nil
-	}
-	if a.taxonomy == nil || a.taxonomy.Sensitivity == nil {
-		return fmt.Errorf("audit: output %q has exclude_labels but taxonomy has no sensitivity config",
-			oe.output.Name())
-	}
-	for label := range oe.excludedLabels {
-		if _, ok := a.taxonomy.Sensitivity.Labels[label]; !ok {
-			return fmt.Errorf("audit: output %q exclude_labels references undefined sensitivity label %q",
-				oe.output.Name(), label)
-		}
-	}
-	return nil
-}
-
-// Accepted trade-off (#509, master-tracker C-23): EnableCategory,
-// DisableCategory, EnableEvent, and DisableEvent all use fmt.Errorf
-// on the validation-miss path, which allocates. These are admin /
-// control-plane operations invoked during startup and occasional
-// runtime reconfiguration — not the hot path. The allocation rate
-// is negligible compared to Audit() and optimising it would obscure
-// the error-message semantics. Hot-path filter state reads
-// (filterState.isEnabled) are separately optimised and do not
-// allocate; see filter.go.
-
-// EnableCategory enables all events in the named category. The
-// category MUST exist in the registered taxonomy. Per-event overrides
-// via [Auditor.DisableEvent] take precedence over category state.
-func (a *Auditor) EnableCategory(category string) error {
-	if a.disabled {
-		return fmt.Errorf("audit: cannot enable category on disabled auditor: %w", ErrDisabled)
-	}
-	// taxonomy is immutable after construction; safe to read without lock.
-	if _, ok := a.taxonomy.Categories[category]; !ok {
-		return fmt.Errorf("audit: unknown category %q", category)
-	}
-	a.filter.enabledCategories.Store(category, true)
-	a.logger.Load().Info("audit: category enabled", "category", category)
-	return nil
-}
-
-// DisableCategory disables all events in the named category. The
-// category MUST exist in the registered taxonomy. Per-event overrides
-// via [Auditor.EnableEvent] take precedence over category state.
-func (a *Auditor) DisableCategory(category string) error {
-	if a.disabled {
-		return fmt.Errorf("audit: cannot disable category on disabled auditor: %w", ErrDisabled)
-	}
-	// taxonomy is immutable after construction; safe to read without lock.
-	if _, ok := a.taxonomy.Categories[category]; !ok {
-		return fmt.Errorf("audit: unknown category %q", category)
-	}
-	a.filter.enabledCategories.Store(category, false)
-	a.logger.Load().Info("audit: category disabled", "category", category)
-	return nil
-}
-
-// EnableEvent enables a specific event type regardless of its
-// category's state. The event type MUST exist in the registered
-// taxonomy. Per-event overrides take precedence over category state.
-func (a *Auditor) EnableEvent(eventType string) error {
-	if a.disabled {
-		return fmt.Errorf("audit: cannot enable event on disabled auditor: %w", ErrDisabled)
-	}
-	// taxonomy is immutable after construction; safe to read without lock.
-	if _, ok := a.taxonomy.Events[eventType]; !ok {
-		return fmt.Errorf("audit: unknown event type %q", eventType)
-	}
-	a.filter.eventOverrides.Store(eventType, true)
-	a.filter.hasEventOverrides.Store(true)
-	a.logger.Load().Info("audit: event enabled", "event_type", eventType)
-	return nil
-}
-
-// DisableEvent disables a specific event type regardless of its
-// category's state. The event type MUST exist in the registered
-// taxonomy. Per-event overrides take precedence over category state.
-func (a *Auditor) DisableEvent(eventType string) error {
-	if a.disabled {
-		return fmt.Errorf("audit: cannot disable event on disabled auditor: %w", ErrDisabled)
-	}
-	// taxonomy is immutable after construction; safe to read without lock.
-	if _, ok := a.taxonomy.Events[eventType]; !ok {
-		return fmt.Errorf("audit: unknown event type %q", eventType)
-	}
-	a.filter.eventOverrides.Store(eventType, false)
-	a.filter.hasEventOverrides.Store(true)
-	a.logger.Load().Info("audit: event disabled", "event_type", eventType)
-	return nil
-}
-
-// SetOutputRoute sets the per-output event route for the named output.
-// The route is validated against the taxonomy; unknown categories or
-// event types return an error. Mixed include/exclude routes return an
-// error. An unknown output name returns an error.
-//
-// SetOutputRoute is safe for concurrent use with event delivery.
-func (a *Auditor) SetOutputRoute(outputName string, route *EventRoute) error {
-	if a.disabled {
-		return fmt.Errorf("audit: cannot set output route on disabled auditor: %w", ErrDisabled)
-	}
-	oe, ok := a.outputsByName[outputName]
-	if !ok {
-		return fmt.Errorf("audit: unknown output %q", outputName)
-	}
-	if err := ValidateEventRoute(route, a.taxonomy); err != nil {
-		return err
-	}
-	oe.setRoute(route)
-	a.logger.Load().Info("audit: output route set", "output", outputName)
-	return nil
-}
-
-// ClearOutputRoute removes the per-output event route for the named
-// output, causing it to receive all globally-enabled events.
-//
-// ClearOutputRoute is safe for concurrent use with event delivery.
-func (a *Auditor) ClearOutputRoute(outputName string) error {
-	oe, ok := a.outputsByName[outputName]
-	if !ok {
-		return fmt.Errorf("audit: unknown output %q", outputName)
-	}
-	oe.setRoute(&EventRoute{})
-	a.logger.Load().Info("audit: output route cleared", "output", outputName)
-	return nil
-}
-
-// OutputRoute returns a copy of the current per-output event route
-// for the named output. An unknown output name returns an error.
-func (a *Auditor) OutputRoute(outputName string) (EventRoute, error) {
-	oe, ok := a.outputsByName[outputName]
-	if !ok {
-		return EventRoute{}, fmt.Errorf("audit: unknown output %q", outputName)
-	}
-	return oe.getRoute(), nil
-}
+// Per-event-type and per-category enable/disable controls
+// ([Auditor.EnableCategory] etc.) are defined in control.go.
+// Per-output route validation and management
+// ([Auditor.SetOutputRoute] etc.) are defined in route.go.
 
 // Handle returns an [EventHandle] for the named event type. Call
 // once at startup (for example during DI wiring), cache the returned
